fix course view bug

+++ course/views.py
@@ -114,6 +114,7 @@ def addbranch(request,id):
     contex = {
         "form": form,
     }
+        
 
     return render(request,"addbranch.html", contex)
 
@@ -132,8 +133,12 @@ def addexam(request,id):
     contex = {
         "form": form,
     }
-
-    return render(request,"addexam.html", contex)
+    
+    if request.user.id == CourseBoss.objects.get(id=id).user_id:
+        return render(request,"addexam.html", contex)
+    else:
+        return redirect("course:detailcourse",id=id)
+    
 
 @login_required(login_url='course/courseexam')
 def courseexam(request,id):
@@ -209,9 +214,11 @@ def trainernotification(request,id):
     contex = {
         "student_applys":student_apply,
     }
-
-
-    return render(request, "trainernotification.html", contex)
+    
+    if request.user.id == Trainer.objects.get(id=id).user_id:
+        return render(request, "trainernotification.html", contex)
+    else:
+        return redirect("course:detailtrainers",id=id)
 
 
 @login_required(login_url='course/trainerapplynotification')
@@ -239,8 +246,10 @@ def coursenotification(request,id):
         "student_infos": student_info,
     }
 
-
-    return render(request,"coursenotification.html",contex)
+    if request.user.id == CourseBoss.objects.get(id=id).user_id:
+        return render(request,"coursenotification.html",contex)
+    else:
+        return redirect("course:detailcourse",id=id)
 
 
 @login_required(login_url='course/courseapplynotification')
@@ -252,6 +261,8 @@ def courseapplynotification(request,id):
         "student_infos": student_info,
 
     }
+    
+    
 
     return render(request,"courseapplynotification.html",contex)
 
@@ -296,8 +307,12 @@ def addphoto(request,id):
         "form": form,
     }
 
-
-    return render(request,"addphoto.html", contex)
+    if request.user.id == CourseBoss.objects.get(id=id).user_id:
+        return render(request,"addphoto.html", contex)
+    else:
+        return redirect("course:detailcourse",id=id)
+    
+    
 
 @login_required(login_url='course/blogdetails')
 def blogdetails(request, id):
@@ -324,8 +339,12 @@ def addservice(request, id):
     contex = {
         "form":form,
     }
-
-    return render (request, "service.html", contex)
+    
+    
+    if request.user.id == CourseBoss.objects.get(id=id).user_id:
+        return render (request, "service.html", contex)
+    else:
+        return redirect("course:detailcourse",id=id)
 
 @login_required(login_url='course/updateservice')
 def updateservice(request, id):
@@ -345,7 +364,10 @@ def updateservice(request, id):
         "form":form,
 
     }
-    return render(request, "updatelessonplan.html",contex)
+    if request.user.id == CourseService.objects.get(id=id).user_id:
+        return render(request, "updatelessonplan.html",contex)
+    else:
+        return redirect("course:detailcourse",id=id)
 
 
 def detailservice(request,id):
@@ -365,8 +387,11 @@ def deleteservice(request,id):
 
     lessonplan = get_object_or_404(CourseService,id=id)
     courseid=CourseService.objects.get(id=id).course_id
-    lessonplan.delete()
-    return redirect ("course:detailcourse", id=courseid)
+    if request.user.id == CourseService.objects.get(id=id).user_id:
+        lessonplan.delete()
+        return redirect ("course:detailcourse", id=courseid)
+    else:
+        return redirect("course:detailcourse",id=id)
 
 
 
@@ -405,8 +430,11 @@ def addevent(request,id):
     contex ={
         "form":form,
     }
-
-    return render (request, "addevent.html", contex)
+    
+    if request.user.id == Trainer.objects.get(id=id).user_id:
+        return render (request, "addevent.html", contex)
+    else:
+        return redirect("course:detailtrainers",id=id)
 
 @login_required(login_url='course/eventdetails')
 def eventdetails(request,id):
@@ -456,8 +484,12 @@ def eventapplynotification(request,id):
 @login_required(login_url='course/deleteevent')
 def deleteevent(request,id):
     event = get_object_or_404(Event,trainer=id)
-    event.delete()
-    return redirect("course:detailtrainers",id=id)
+    if request.user.id == Trainer.objects.get(id=id).user_id:
+        event.delete()
+        return redirect("course:detailtrainers",id=id)
+    else:
+        return redirect("course:detailtrainers",id=id)
+        
 
 
 
@@ -480,8 +512,11 @@ def updateevent(request,id):
         "form":form,
 
     }
-
-    return render(request, "updateevent.html",contex)
+    if request.user.id == Trainer.objects.get(id=id).user_id:
+        return render(request, "updateevent.html",contex)
+    else:
+        return redirect ("course:detailtrainers",id=id)
+        
 
 
 
@@ -535,8 +570,12 @@ def updatecourse(request,id):
         "form":form,
 
     }
-
-    return render(request, "updatecourse.html",contex)
+    
+    if request.user.id == CourseBoss.objects.get(id=id).user_id:
+        return render(request, "updatecourse.html",contex)
+    else:
+        return redirect("course:detailcourse",id=id)
+        
 
 
 
@@ -559,7 +598,11 @@ def updatetrainer(request,id):
 
     }
 
-    return render(request, "updatetrainer.html",contex)
+    if request.user.id == Trainer.objects.get(id=id).user_id:
+        return render(request, "updatetrainer.html",contex)
+    else:
+        return redirect("course:detailtrainers",id=id)
+        
 
 
 @login_required(login_url='course/applyexam')
@@ -593,16 +636,24 @@ def applyexam(request,id):
 @login_required(login_url='course/deletecourse')
 def deletecourse(request,id):
     course = get_object_or_404(CourseBoss,id=id)
-    course.delete()
-    return redirect("course:courses")
+    if request.user.id == CourseBoss.objects.get(id=id).user_id:
+        course.delete()
+        return redirect("course:courses")
+    else:
+        return redirect("course:detailcourse",id=id)
+        
 
 
 
 @login_required(login_url='course/deletetrainer')
 def deletetrainer(request, id):
     trainer = get_object_or_404(Trainer,id=id)
-    trainer.delete()
-    return redirect("course:courses")
+    if request.user.id == Trainer.objects.get(id=id).user_id:
+        trainer.delete()
+        return redirect("course:courses")
+    else:
+        return redirect("course:detailtrainers",id=id)
+        
 
 
 
@@ -623,16 +674,25 @@ def updatebranch(request,id):
         "branch":branch
 
     }
-
-    return render(request, "updatebranch.html",contex)
+   
+    if request.user.id == CourseBoss.objects.get(id=courseid).user_id:
+        return render(request, "updatebranch.html",contex)
+    else:
+        return redirect("course:detailcourse", id=courseid)
+        
+        
 
 
 @login_required(login_url='course/deletebranch')
 def deletebranch(request, id):
     branch = get_object_or_404(Branchs,id=id)
     courseid = Branchs.objects.get(id=id).branch_id
-    branch.delete()
-    return redirect("course:detailcourse", id=courseid)
+    if request.user.id == CourseBoss.objects.get(id=courseid).user_id:
+        branch.delete()
+        return redirect("course:detailcourse", id=courseid)
+    else:
+        return redirect("course:detailcourse", id=courseid)
+        
 
 
 @login_required(login_url='course/comment')
@@ -665,43 +725,59 @@ def gallery(request,id):
 def confirm(request, id):
     status = CourseApply.objects.filter(id=id).first()
     courseid = CourseApply.objects.get(id=id).course_id
-    status.confirm = 1
-    status.pending = 0
-    status.cancel = 0
-    status.save()
-    return redirect("course:coursenotification",id=courseid)
+    if request.user.id == CourseBoss.objects.get(id=courseid).user_id:
+        status.confirm = 1
+        status.pending = 0
+        status.cancel = 0
+        status.save()
+        return redirect("course:coursenotification",id=courseid)
+    else:
+        return redirect("course:detailcourse",id=id)
+        
 
 
 @login_required(login_url='course/confirmtrainer')
 def confirmtrainer(request, id):
     status = TrainerApply.objects.filter(id=id).first()
     trainerid = TrainerApply.objects.get(id=id).trainer_id
-    status.confirm = 1
-    status.pending = 0
-    status.cancel = 0
-    status.save()
-    return redirect("course:trainernotification",id=trainerid)
+    if request.user.id == Trainer.objects.get(id=trainerid).user_id:
+        status.confirm = 1
+        status.pending = 0
+        status.cancel = 0
+        status.save()
+        return redirect("course:trainernotification",id=trainerid)
+    else:
+        return redirect("course:detailtrainers",id=trainerid)
 
 
 @login_required(login_url='course/canceltrainer')
 def canceltrainer(request, id):
     status = TrainerApply.objects.filter(id=id).first()
-    status.confirm = 0
-    status.pending = 0
-    status.cancel = 1
-    status.save()
-    return redirect("course:trainernotification",id=id)
+    trainerid = TrainerApply.objects.get(id=id).trainer_id
+    if request.user.id == Trainer.objects.get(id=trainerid).user_id:
+        status.confirm = 0
+        status.pending = 0
+        status.cancel = 1
+        status.save()
+        return redirect("course:trainernotification",id=id)
+    else:
+        return redirect("course:detailtrainers",id=trainerid)
+        
 
 
 @login_required(login_url='course/cancel')
 def cancel(request,id):
     courseid = CourseApply.objects.get(id=id).course_id
     status = CourseApply.objects.filter(id=id).first()
-    status.confirm = 0
-    status.pending = 0
-    status.cancel = 1
-    status.save()
-    return redirect("course:coursenotification",id=courseid)
+    if request.user.id == CourseApply.objects.get(id=courseid).user_id:
+        status.confirm = 0
+        status.pending = 0
+        status.cancel = 1
+        status.save()
+        return redirect("course:coursenotification",id=courseid)
+    else:
+        return redirect("course:detailcourse",id=courseid)
+        
 
 
 def createblog(request):
@@ -722,7 +798,7 @@ def createblog(request):
         "form": form,
 
     }
-
+    
     return render(request, "addblog.html", contex)
 
 def articles(request):
@@ -752,12 +828,19 @@ def updatearticle(request,id):
         "form": form,
 
     }
-    return render (request,"updatearticle.html", contex)
+    
+    if request.user.id == get_object_or_404(CreateBlog,user_id=id):
+        return render (request,"updatearticle.html", contex)
+    else:
+        return redirect("course:articles")
 
 def deletearticle(request,id):
     blog = get_object_or_404(CreateBlog,user_id=id)
-    blog.delete()
-    return redirect("course:articles")
+    if request.user.id == get_object_or_404(CreateBlog,user_id=id):
+        blog.delete()
+        return redirect("course:articles")
+    else:
+        return redirect("course:articles")
 
 
 
@@ -801,7 +884,11 @@ def coursemessage(request,id):
         "form":form
 
     }
-    return render (request,"messages.html", contex)
+    course_id = CourseApply.objects.get(id=id).course_id
+    if request.user.id == CourseBoss.objects.get(id=course_id).user_id:
+        return render (request,"messages.html", contex)
+    else:
+        return redirect("course:detailcourse", id=course_id)
 
 
 def usermessage(request,id):
@@ -825,7 +912,12 @@ def usermessage(request,id):
         "form":form
 
     }
-    return render (request,"messages.html", contex)
+    userid = CourseApply.objects.get(id=id).user_id
+    course_id = CourseApply.objects.get(id=id).course_id
+    if request.user.id == userid:
+        return render (request,"messages.html", contex)
+    else:
+        return redirect("course:detailcourse", id=course_id)
